+++ packages/common/src/HdWallet.ts
@@ -2,7 +2,12 @@ import * as wasm from '@concordium/rust-bindings';
 import { mnemonicToSeedSync, validateMnemonic } from '@scure/bip39';
 import { wordlist } from '@scure/bip39/wordlists/english';
 import { Buffer } from 'buffer/';
-import { AttributesKeys, Network, CryptographicParameters } from './types';
+import {
+    AttributesKeys,
+    Network,
+    CryptographicParameters,
+    ContractAddress,
+} from './types';
 import { isHex } from './util';
 
 /**
@@ -153,12 +158,15 @@ export class ConcordiumHdWallet {
     }
 
     getVerifiableCredentialSigningKey(
+        issuer: ContractAddress,
         verifiableCredentialIndex: number
     ): Buffer {
         return Buffer.from(
             wasm.getVerifiableCredentialSigningKey(
                 this.seedAsHex,
                 this.network,
+                issuer.index,
+                issuer.subindex,
                 verifiableCredentialIndex
             ),
             'hex'
@@ -166,26 +174,26 @@ export class ConcordiumHdWallet {
     }
 
     getVerifiableCredentialPublicKey(
+        issuer: ContractAddress,
         verifiableCredentialIndex: number
     ): Buffer {
         return Buffer.from(
             wasm.getVerifiableCredentialPublicKey(
                 this.seedAsHex,
                 this.network,
+                issuer.index,
+                issuer.subindex,
                 verifiableCredentialIndex
             ),
             'hex'
         );
     }
 
-    getVerifiableCredentialEncryptionKey(
-        verifiableCredentialIndex: number
-    ): Buffer {
+    getVerifiableCredentialBackupEncryptionKey(): Buffer {
         return Buffer.from(
-            wasm.getVerifiableCredentialEncryptionKey(
+            wasm.getVerifiableCredentialBackupEncryptionKey(
                 this.seedAsHex,
-                this.network,
-                verifiableCredentialIndex
+                this.network
             ),
             'hex'
         );